Prefer the use of uXX and sXX types in `JKRDvdArchive` and `JKRArchive`

+++ libs/JSystem/include/JSystem/JKernel/JKRArchive.hpp
@@ -103,14 +103,14 @@ public:
     };
 
     JKRArchive();
-    JKRArchive(long, EMountMode);
+    JKRArchive(s32, EMountMode);
     virtual ~JKRArchive();
 
     virtual bool becomeCurrent(const char*);
     virtual void* getResource(const char*);
-    virtual void* getResource(unsigned long, const char*);
-    virtual u32 readResource(void*, unsigned long, const char*);
-    virtual u32 readResource(void*, unsigned long, unsigned long, const char*);
+    virtual void* getResource(u32, const char*);
+    virtual u32 readResource(void*, u32, const char*);
+    virtual u32 readResource(void*, u32, u32, const char*);
     virtual void removeResourceAll();
     virtual bool removeResource(void*);
     virtual bool detachResource(void*);
@@ -118,30 +118,30 @@ public:
     virtual s32 countFile(const char*) const;
     JKRArcFinder* getFirstFile(const char*) const;
     virtual s32 getExpandedResSize(const void*) const;
-    virtual void* fetchResource(SDIFileEntry*, unsigned long*) = 0;
-    virtual void* fetchResource(void*, unsigned long, SDIFileEntry*, unsigned long*) = 0;
-    virtual void setExpandSize(SDIFileEntry*, unsigned long);
+    virtual void* fetchResource(SDIFileEntry*, u32*) = 0;
+    virtual void* fetchResource(void*, u32, SDIFileEntry*, u32*) = 0;
+    virtual void setExpandSize(SDIFileEntry*, u32);
     virtual u32 getExpandSize(SDIFileEntry*) const;
 
-    static JKRArchive* check_mount_already(long);
-    static JKRArchive* check_mount_already(long, JKRHeap*);
+    static JKRArchive* check_mount_already(s32);
+    static JKRArchive* check_mount_already(s32, JKRHeap*);
     static void mount(const char*, EMountMode, JKRHeap*, EMountDirection);
-    static JKRArchive* mount(long, EMountMode, JKRHeap*, EMountDirection);
-    bool getDirEntry(SDirEntry*, unsigned long) const;
-    void* getIdxResource(unsigned long);
-    void* getResource(unsigned short);
-    u32 readResource(void*, unsigned long, unsigned short);
+    static JKRArchive* mount(s32, EMountMode, JKRHeap*, EMountDirection);
+    bool getDirEntry(SDirEntry*, u32) const;
+    void* getIdxResource(u32);
+    void* getResource(u16);
+    u32 readResource(void*, u32, u16);
     u32 countResource() const;
-    u32 getFileAttribute(unsigned long) const;
-    bool isSameName(CArcName&, unsigned long, unsigned short) const;
-    SDIDirEntry* findResType(unsigned long) const;
-    SDIDirEntry* findDirectory(const char*, unsigned long) const;
-    SDIFileEntry* findTypeResource(unsigned long, const char*) const;
-    SDIFileEntry* findFsResource(const char*, unsigned long) const;
-    SDIFileEntry* findIdxResource(unsigned long) const;
+    u32 getFileAttribute(u32) const;
+    bool isSameName(CArcName&, u32, u16) const;
+    SDIDirEntry* findResType(u32) const;
+    SDIDirEntry* findDirectory(const char*, u32) const;
+    SDIFileEntry* findTypeResource(u32, const char*) const;
+    SDIFileEntry* findFsResource(const char*, u32) const;
+    SDIFileEntry* findIdxResource(u32) const;
     SDIFileEntry* findNameResource(const char*) const;
     SDIFileEntry* findPtrResource(const void*) const;
-    SDIFileEntry* findIdResource(unsigned short) const;
+    SDIFileEntry* findIdResource(u16) const;
 
     static u32 sCurrentDirIndex;  // 0x806B7148
